requirements.txt
aiogram==3.18.0
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
aiohttp==3.11.13
